fix: renaming component
