+++ src/pages/affiliates/index.tsx
@@ -15,7 +15,7 @@ import { AffiliatesText } from '@/assets/types/affiliates';
 import { FooterText, MetatagsDescription, NavigationBarText } from '@/assets/types/layout/types';
 import { IMPACT_LOGIN_USER, SIGNUP_DRIVE_WEB } from '@/constants';
 import { TextAndCardsGroupColumnSection } from '@/components/shared/components/TextAndCardsGroupColumnSection';
-import { ComponentsInARowSection } from '@/components/shared/components/ComoponentsInARowSection';
+import { ComponentsInARowSection } from '@/components/shared/components/ComponentsInARowSection';
 import Button from '@/components/shared/Button';
 import { getImage } from '@/lib/getImage';
 import Image from 'next/image';
